Remove __init__.py editing step
The initial commit from cookiecutter is already contained `from .functions import *` and all of the function inside `functions.py` already available. In MolSSI May workshop recording it also clearly shows that `from .functions import *` is the default, while the `__init__.py` editing is only for the sake of  demonstrating the various alternative of package import.