Fixed Github.com hyperlink in setup.md and more.
Previous page's hyperlink lead to "github.coms" for the "Create GitHub Account" section which doesn't exist. For the lines "A more complete list of possible editors is available:",  and "Download the starting workshop materials:" I've changed the hyperlink's text to something other than "here" which is considered bad practice in an HTML book I read in the past. The directory structure section doesn't mention the ".DS_Store" file which is included in several directories of the starting_material.zip archive, so I've mentioned that those files should be ignored.